requirements.txt
flask>=3.0
google-genai>=1.68.0
pillow>=12.0
python-dotenv>=1.0
